Global define NOMINMAX. Fix enable_vscroll_bar in Win32.

+++ test/sources/WindowTest.cpp
@@ -512,7 +512,7 @@ int main(int argc, char* argv[]) {
   list5.set_visible();
 
   vscroll.create(main, core::rectangle(700, 50, 16, 250));
-  vscroll.set_max(list1.get_count() * list1.get_item_size().height() - list1.size().height());
+  vscroll.set_max((int)list1.get_count() * list1.get_item_size().height() - list1.size().height());
   vscroll.set_step(list1.get_item_size().height());
   vscroll.set_visible();
